Use TOFLength from the data file
It seems like run 2013 processing started with wrong a carbon travel
distance value of 18cm. The input data file format at that time was already
providing that value. Correct value for run 2013 seems to be 19.5 which
is also the value I found in the input files.

This commit enables using TOFLength value from the data files.