CachingLogReader: change database path
Before that CachingLogReader would store it's cache in it's current
working directory which in our case is often stored on NFS mount.
Because SQLite heavily relies on file locks this seems to produce some
databse locks.

The proposed workaround is to move SQLite DB file to /tmp which should
provide more decent file locking support.