CnipolAsymHists: TDC cut to fix broken alphas
Some alpha measurements (for example 18706.206) get taken while there is
still some beam in the ring. We can still save some of the information
by looking at the event distribution by TDC. The bad events manifest
themselves as a big bump in that distribution. Comparing the event count
in TDC bin with the baseline of the distribution we can detect such
events and remove them from alpha peak histograms.