CnipolAlphaHists: adjust ADC bins
This sets histograms to display ADC bin values in the range from 0 to
219. Higher ADC value bins(220,221) are WFD overflow and they now will
go to overflow bin of the histogram. That way overflow bins won't be
mistaken for alpha peaks.

Actually overflow bin is supposed to be ADC=221, but I often see many
overflowish events go to 220. Igor claims that there is a jitter due to
pedestals variation.

18701.201.alpha0 is a good example of such overflow.