implement some "new system" streamers
This should allow to read v3 of MeasInfo and v1 of AnaMeasResult as well
as automatically provide schema migrations in the future. These two
classes are important to be able to migrate because they are referenced
by EventConfig and they do contain useful information.

The change was implemented following instructions for the Case B in
https://root.cern.ch/root/SchemaEvolution.html

AnaMeasResult+ pragma in LinkDef was overriden by the one below, so we
remove it as it has no effect.