New system for getting run period parameters
Right now parameters such as alpha run, number of alpha sources,
disabled channels mask and banana plot cuts are stored in MySQL in a
"run_period" table. Even though MySQL worked very well for us
for storing measurements and fill results, it doesn't really fit
for this kind of input data.

 * This leaves no way to do meaningfull processing without MySQL
   database up. --use-db flag and available DB server become a
   requirement.

 * This doesn't scale very well for multiple users. When modifying run
   period constants users will interfere unless each has a personal DB
   to play with. In later case it becomes cumbersome to do
   synchronization between theese databases.

 * Changes to MySQL database are hard to track and document. Even though
   the values are appearing in runconfig.php files there are no tools to
   track those as well.

New system introduces a POD type RunPeriod and stores RunPeriods as
a C-style array. Such approach allows to use git for tracking changes in
run period parameters. The neglection of using text file formats (using
libconfig or whatnot) is intentional - this is done to prevent end user
from being able to change the parameters without changing fAsymVersion.

RunPeriod is for the moment implemented in C. It's probably a bad idea
to mix C and C++, but I really wanted to use designated initializer
lists feature for RunPeriod.inc which is not even in C++11, but present
in C99 and compiles by C compiler on our RHEL6 system.