AlphaCalibrator: increase blur rate in peak finder
This is done without lowering the minimal first peak height. Because of
that effective minimal peak height raises, but that's ok.
The effect of this change is to help finding Am peak. Extended bluring
allows for filtering statistical fluctuations so that Gd peak has less
chance to shoot high enough be misidentified as highest (Am) peak.