CachingLogReader: prevent hanging database locks
This fixes a SQLite cache database deadlock that happened after
unsuccessful destruction of CachingLogReader object. The next created
object would be unable to open the database for writing. The problem
occured when masym was used.

The solution was to finalize all prepared statements so that
sqlite3_close() can do it's job. The check for it's return code was also
added.

The bug was initially reported by Grant Webb.